+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `/reset` page for clearing all data
+
 ## [v1.1.2] - 2020-10-02
 
 ### Added